He añadido el resto de carpetas al proyecto y modificado distintas cosas como el pageable correctamente... Falta la creacion de DTO y seguir añadiendo cosas pero he conseguido que corra y tirar consultas en el thunderclient

+++ src/main/java/com/example/proyectoServidor/repository/DeveloperRepository.java
@@ -6,7 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.example.proyectoServidor.model.Developer;
 
 @Repository
-public interface DeveloperRepository extends JpaRepository<Developer, Long> {
+public interface DeveloperRepository extends JpaRepository<Developer, Integer> {
 
     
 }
+// Accede y maneja la base de datos
