+++ src/main/java/com/example/proyectoServidor/repository/TechnologyRepository.java
@@ -6,7 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.example.proyectoServidor.model.Technology;
 
 @Repository
-public interface TechnologyRepository extends JpaRepository<Technology, Long> {
+public interface TechnologyRepository extends JpaRepository<Technology, Integer> {
 
     
 } 
+// Accede y maneja la base de datos
